Replace the chiptype-to-name lookup table with a big "if" block.
The mapping in a lookup table is much clearer to read, but it has
one disadvantage: if the used romloader plugin does not define one
of the chip types, the table can not be parsed. The big "if" block
is not so pretty to look at, but it is much more robust.