Replace the bus and unit tables with a unit description.
The new unit description table maps a bus-unit-chipselect tuple to an ASCII
identifier (the ID) and a set of flags (currently unused). The ID is much
easier to remember for an end user than the old bus-unit-chipselect tuple.
Furthermore the mapping provides all available chip selects, which was missing
in the old tables. Now it is possible for a tool like the CLI flasher to
provide a complete list of available flash units to the user.