Remove the term "wonderful feelings" as it does not fit.
The name WFP has it's origin in the german construction "Wohlfühlpaket".
A rough translation could be "feel-good package". It should describe the
feeling of relief that the old flash process for the netX90 can now be done
with one simple command instead of having to flash about 8 files separately
and having to guess the offset address for some of them.
The name "wonderful feelings" did not fit this intention very well.